Moved db_settings.py from settings folder to tg_bot_float_db_app folder. Added db_app_variables.env file. Fixed BotDbRefresherService
