+++ src/tg_bot_float_db_app/bot_db_exception.py
