+++ src/tg_bot_float_db_app/database/db_factory.py
@@ -2,7 +2,7 @@ from sqlalchemy.ext import asyncio as async_alchemy
 
 from tg_bot_float_db_app.database.bot_db_service_factory import BotDbServiceFactory
 from tg_bot_float_db_app.database.bot_db_creator import BotDbCreator
-from settings.db_settings import DBSettings
+from tg_bot_float_db_app.db_settings import DBSettings
 
 
 class BotDbFactory:
