+++ src/tg_bot_float_db_app/database/services/quality_service.py
@@ -11,8 +11,8 @@ from tg_bot_float_db_app.database.models.quality_model import QualityModel
 from tg_bot_float_db_app.database.models.relation_model import RelationModel
 from tg_bot_float_db_app.database.models.skin_model import SkinModel
 from tg_bot_float_db_app.database.models.weapon_model import WeaponModel
-from tg_bot_float_db_app.misc.bot_db_exception import BotDbException
-from tg_bot_float_db_app.misc.router_constants import (
+from tg_bot_float_db_app.bot_db_exception import BotDbException
+from tg_bot_float_db_app.api.routers.router_constants import (
     ENTITY_FOUND_ERROR_MSG,
     ENTITY_NOT_FOUND_ERROR_MSG,
     NONE_FIELD_IN_ENTITY_ERROR_MSG,
@@ -192,9 +192,13 @@ class QualityService:
             )
         await self._session.commit()
 
-    async def get_all(self) -> Page[QualityDTO]:
+    async def get_all_paginated(self) -> Page[QualityDTO]:
         select_stmt = select(QualityModel)
         return await paginate(self._session, select_stmt)
+
+    async def get_all(self) -> ScalarResult[QualityModel]:
+        select_stmt = select(QualityModel)
+        return await self._session.scalars(select_stmt)
 
     async def get_many_by_weapon_skin_name(self, weapon_name: str, skin_name: str):
         stmt = (
