+++ src/tg_bot_float_db_app/database/services/weapon_service.py
@@ -10,8 +10,8 @@ from fastapi_pagination.ext.sqlalchemy import paginate
 from tg_bot_float_db_app.database.models.relation_model import RelationModel
 from tg_bot_float_db_app.database.models.skin_model import SkinModel
 from tg_bot_float_db_app.database.models.weapon_model import WeaponModel
-from tg_bot_float_db_app.misc.bot_db_exception import BotDbException
-from tg_bot_float_db_app.misc.router_constants import (
+from tg_bot_float_db_app.bot_db_exception import BotDbException
+from tg_bot_float_db_app.api.routers.router_constants import (
     ENTITY_FOUND_ERROR_MSG,
     ENTITY_NOT_FOUND_ERROR_MSG,
     NONE_FIELD_IN_ENTITY_ERROR_MSG,
@@ -193,9 +193,13 @@ class WeaponService:
             )
         await self._session.commit()
 
-    async def get_all(self) -> Page[WeaponDTO]:
-        stmt = select(WeaponModel)
-        return await paginate(self._session, stmt)
+    async def get_all_paginated(self) -> Page[WeaponDTO]:
+        select_stmt = select(WeaponModel)
+        return await paginate(self._session, select_stmt)
+
+    async def get_all(self) -> ScalarResult[WeaponModel]:
+        select_stmt = select(WeaponModel)
+        return await self._session.scalars(select_stmt)
 
     async def get_many_by_skin_name(self, skin_name: str) -> ScalarResult[WeaponModel]:
         stmt = (
