+++ src/tg_bot_float_db_app/db_app_variables.env
@@ -1,0 +1,2 @@
+url="postgresql+asyncpg://classified:rv9up0ax@192.168.0.200:5432/tg_bot_float_db"
+
