--- src/tg_bot_float_db_app/misc/exceptions.py
@@ -1,3 +1,0 @@
-class BotDbException(Exception):
-    def __init__(self, msg: str):
-        self.msg = msg
